Msg file to communicate
Hello i have figured out how to have two SSID's after ORBI boots up, there are a few steps that need to be executed, the files on the rom if can be edited 
it would be hte easy solution. Since that is not possible probably you can help me here with your project.

Let me know if you are interested in getting hte steps to get two SSID's one for 2.4 and one for 5 Ghz. The new firmware does not 
allow you to have two SSID's but these steps should resolve the issue

I am new to Git Hub , did not know how to comunicate so i am creatig a file here

THank you